split crate in 2
* statement contains only traits and simple wrapper types
* statement_validation contains code for validating the contract
  of the wrapper types
