+++ statement_validation/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "r2c2_statement"
+name = "r2c2_statement_validation"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -9,6 +9,9 @@ license-file.workspace = true
 keywords.workspace = true
 
 [dependencies]
+r2c2_statement.workspace = true
+regex = "1.11.1"
+
 langtag = { version = "0.4.0", optional = true }
 oxrdf = { version = "0.2.4", optional = true, features = ["rdf-star"] }
 rdf-types = { version = "0.22.5", optional = true }
