+++ statement_validation/src/lib.rs
@@ -1,0 +1,8 @@
+//! I extends the utility types of [`r2c2_statement`]
+//! with validating constructors, as a convenience for implementers.
+#![deny(missing_docs)]
+
+mod _iri;
+pub use _iri::*;
+mod _language_tag;
+pub use _language_tag::*;
